@@ -16,7 +16,7 @@ dependencies:
   animate_do: ^2.1.0
   auto_animated: ^3.0.1
   bloc: ^8.0.3
-  carrier_info: ^2.0.4
+  carrier_info: ^2.0.5
   cupertino_icons: ^1.0.4
   elegant_notification: ^1.5.2
   enum_to_string: ^2.0.1
@@ -25,13 +25,13 @@ dependencies:
   equatable: ^2.0.3
   feedback: ^2.4.1
   feedback_sentry: ^2.3.0
-  flex_color_scheme: ^4.2.0
+  flex_color_scheme: ^5.0.0
   flutter_bloc: ^8.0.1
   flutter_svg: ^1.0.3
   flutter_platform_widgets: ^1.20.0
   font_awesome_flutter: ^10.1.0
-  google_fonts: ^2.3.1
-  google_maps_flutter: ^2.1.3
+  google_fonts: ^2.3.2
+  google_maps_flutter: ^2.1.4
   google_mobile_ads: ^1.2.0
   hive: ^2.1.0
   hive_flutter: ^1.1.0
@@ -39,8 +39,8 @@ dependencies:
   hydrated_bloc: ^8.1.0
   in_app_purchase: ^3.0.2
   introduction_screen: ^3.0.2
-  json_serializable: ^6.1.5
-  json_annotation: ^4.4.0
+  json_serializable: ^6.2.0
+  json_annotation: ^4.5.0
   lan_scanner: ^3.4.1
   network_info_plus: ^2.1.3
   modal_bottom_sheet: ^2.0.1
@@ -48,16 +48,16 @@ dependencies:
   path_provider: ^2.0.9
   permission_handler: ^9.2.0
   rxdart: ^0.27.3
-  sentry_flutter: ^6.4.0
+  sentry_flutter: ^6.5.1
   styled_widget: ^0.4.0+3
-  url_launcher: ^6.0.20
+  url_launcher: ^6.1.0
   wake_on_lan: ^3.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  build_runner: ^2.1.8
+  build_runner: ^2.1.10
   dart_code_metrics: ^4.14.0
   flutter_launcher_icons_maker: ^0.10.2
   import_sorter: ^4.6.0
